Add flag to parse user overrides for code words
I had to add the ability to get the mappings from a PhoneticConverter
since that's where we define overrides (not on SpellingAlphabet's
directly) and we also need the `--dump- alphabet` flag to display them
accordingly.

The one caveat to the current parsing code is that it won't let
users override a comma (',') or equal sign ('=') since it uses those
characters to parse the string argument and does not support escaping.
It could be addressed eventually, but would make the code more complex.

To improve the user experience and ensure consistency, we're also
normalizing the conversion map values to always be pascal case. The
convert_case crate we're using properly handles a number of different
forms of input.

See:
- https://crates.io/crates/convert_case/