Keep user input at end of error output
This makes it a bit easier to read and understand when the user input
that is causing an error isn't sandwiched in the middle of the error
output message. I tried creating custom clap errors, but I actually
didn't like the output as much and I don't mind having a distinction
between output for clap-related errors and errors with my own parsing
code.