Expand the README with details on CLI usage
Writing out the details for the environment variables that expect
a boolean value made me realize that we also don't need the full
functionality of clap's BoolishValueParser. Since the default is already
false, we can assume everything that's not "falsey" should be considered
true rather than throwing an error to the end user. That should more
closely match their intension and improve the user experience.

See:
- https://docs.rs/clap/latest/clap/builder/struct.FalseyValueParser.html