Tweak documentation surrounding test commands
Since doctests are slow, we'll split out the steps and just show `cargo
nextest run` as the normal workflow. Our `cargo xtask test` wrapper is
still exhaustive though.

I'm also removing the Insta configuration file, as eventually the auto
setting will pick up Nextest as the preferred default, and I couldn't
get it to pay attention to the config file anyway. No need to keep it
around for the moment.