Copy build script outputs to dist directory via xtask
This is so much more complicated than it needs to be. The complexity
comes from the build.rs OUT_DIR being randomly generated, and there's
not an easy way to move files out of there or even reference where the
output directory is. Unfortunately, putting things like the completions
files and man pages in the OUT_DIR is the "proper" way to do things, and
other projects have all sorts of ugly workarounds.

There is an unstable nightly flag `--out-dir` that I thought would
solve the problem, but it only copies the build binary and not the build
script outputs. The workaround here is to parse the json cargo messages
directly to grab the OUT_DIR, and then recursively copy any asset files
ourselves.

See:
- https://github.com/rust-lang/cargo/issues/6790
- https://github.com/benesch/cargo-out-dir/blob/master/src/main.rs