Use --generate flag on spellout binary for printing asset files
The mechanism of using the build script outputs is very clunky due to
the randomized OUT_DIR, and it's also not very discoverable. The main
advantage of keeping the build.rs script is that our `xtask dist` script
doesn't need deep integration with the binary, but either way we're
relying on conventions here. Although it's common to see projects use
build.rs for these asset files, it's really meant to be a mechanism to
compile third-party non-Rust code and other tasks related to the build
itself.

Moving asset file generation into our xtask script directly is also not
a great option, as we need access to the defined clap Command, which I
fear would mean even more direct integration between the binary crate
and our xtasks, which I'd like to try and keep as generic and reusable
as possible.

Finally, having a flag directly on the binary can be useful for people
that want to source a completions file directly from their shell startup
scripts. The completions would always be in sync with the current
version, and the user could take advantage of them even if their package
manager doesn't support installing the project directly. This is also a
pattern I've observed in other projects, like `jj`.

See:
- https://users.rust-lang.org/t/pre-rfc-generating-assets-products-during-cargo-build/32824
- https://github.com/martinvonz/jj/blob/78a64edb22364e01e27836792b896f59d214e39e/src/commands/mod.rs#L908-L914