Remove redundant acronym text
This was actually causing a bug in the Fish shell completion, because
fish uses parentheses for subshells, so when you would try to tab-
complete an alphabet, it was launching a subshell and trying to run the
`NATO` command, for example. That does not exist, so it would spit out
an error and not complete as expected.

Since the option argument is itself already the acronym, we might as
well remove the redundant version to fix the completion script.