Add current git hash to the version output
This build metadata can be useful to track down the exact commit used to
compile the application. It will also check if the repository was dirty
(has modifications) and append that information accordingly.

Using 12 characters to abbreviate the Git hash should be more than
sufficient to be unambiguous, especially considering we're still using
the current cargo package version as well.

See:
- https://semver.org/#spec-item-10
- https://blog.cuviper.com/2013/11/10/how-short-can-git-abbreviate/