Use version output similar to rustc
Since we're not actually publishing the crate with the build metadata,
we don't need to be strict with the semver formatting; it's only
additional information in the application's output. I've retained the
abbreviated hash length of 12 as well as the `.dirty` flag, because I
think that's useful.

This also exports the full commit hash and commit date into the
environment, so we could potentially add a `--version --verbose` output,
similar to rustc as well.