Explain why we fallback to CARGO_PKG_VERSION
Many packagers will prefer to grab a snapshot source tarball rather
than use Git directly to clone a project's source code when packaging
up a release for distribution. Since our build script uses Git to inject
metadata into the environment, that won't work if we're in compiling
from a source tarball and thus our custom version won't exist. In that
case, we want to fallback to Cargo's default behavior. This already
works, I just wanted to document it for posterity.