Configure workspace for cargo-release compatibility
Using cargo-release will help eliminate mistakes when preparing to cut a
new release. It will still take a few steps, but we can write jobs to be
triggered manually that will keep the process consistent with no need to
run them locally.

I've made the xtask crate explicitly internal and it will be ignored
by cargo-release. Since we're managing spellabc as a public library, it
should maintain its own versioning and CHANGELOG.

See:
- https://github.com/crate-ci/cargo-release/
- https://matklad.github.io/2021/08/22/large-rust-workspaces.html#Smaller-Tips