Allow the tagged GITHUB_REF when publishing to crates.io
These tags all exist on the main branch, as desired, but when we trigger
a job from the tag creation, we end up in a detached HEAD state. As
long as we're on the tag that we're publishing, that is okay and we
can override things to allow generating the package and publishing it
to crates.io.