Do not rely on implicit use of cross
Trying to conditionally set the CARGO environment variable is too
complex due to the xtask wrapper and the environment within an xshell
Shell, so instead we'll just add an explicit option flag to select the
command. I believe it's only necessary for Linux when not targetting the
native host triple.