Check out full Git history when publishing crates
Similar to the other cargo release jobs, I think some of the timing
issues I'm observing is it trying to look through history and verify
tags and other repository metadata over the network. With a shallow
clone, it seems to struggle for a bit before timing out and continuing
on. I think this should help.